Similar refactoring for adventures
